+++ skyflow/common/common.go
@@ -3,6 +3,8 @@ Copyright (c) 2022 Skyflow, Inc.
 */
 package common
 
+import "context"
+
 type ResponseBody map[string]interface{}
 type TokenProvider func() (string, error)
 
@@ -29,6 +31,14 @@ const (
 	REDACTED   RedactionType = "REDACTED"
 )
 
+type BYOT string
+
+const (
+	DISABLE       BYOT = "DISABLE"
+	ENABLE        BYOT = "ENABLE"
+	ENABLE_STRICT BYOT = "ENABLE_STRICT"
+)
+
 type ConnectionConfig struct {
 	ConnectionURL string
 	MethodName    RequestMethod
@@ -39,12 +49,25 @@ type ConnectionConfig struct {
 }
 
 type InsertOptions struct {
-	Tokens bool
-	Upsert []UpsertOptions
+	Tokens          bool
+	Upsert          []UpsertOptions
+	Context         context.Context
+	ContinueOnError bool
+	Byot            BYOT
 }
 type GetOptions struct {
 	Tokens bool
 }
+
+type DetokenizeOptions struct {
+	Context         context.Context
+	ContinueOnError bool
+}
+
+type GetByIdOptions struct {
+	Context context.Context
+}
+
 type UpsertOptions struct {
 	Table  string
 	Column string
@@ -58,11 +81,16 @@ type Configuration struct {
 
 type InsertRecords struct {
 	Records []InsertRecord
+	Errors  []InsertError
 }
-
+type InsertError struct {
+	Error InsertResponseError
+}
 type InsertRecord struct {
-	Table  string
-	Fields map[string]interface{}
+	RequestIndex int `json:"request_index"`
+	Table        string
+	Fields       map[string]interface{}
+	Tokens       map[string]interface{}
 }
 
 type DetokenizeInput struct {
@@ -89,11 +117,15 @@ type DetokenizeError struct {
 	Error ResponseError
 }
 
+type InsertResponseError struct {
+	RequestIndex int `json:"request_index"`
+	Code         string
+	Description  string
+}
 type ResponseError struct {
 	Code        string
 	Description string
 }
-
 type GetByIdInput struct {
 	Records []SkyflowIdRecord
 }
@@ -140,7 +172,7 @@ type GetRecord struct {
 type GetError struct {
 	Ids          []string
 	ColumnValues []string
-	ColumnName    string
+	ColumnName   string
 	Error        ResponseError
 }
 
@@ -155,4 +187,4 @@ const (
 )
 
 const sdk_name = "skyflow-go"
-const sdk_version = "1.6.0"
+const sdk_version = "1.9.0"