+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.9.0] - 2023-09-29
+### Added
+- Added BYOT strict modes in Insert Method.
+
+## [1.8.1] - 2023-09-08
+### Added
+- Added request index in response in Insert Method.
+
+## [1.8.0] - 2023-09-01
+### Added
+- Support for Bulk request with Continue on Error in Detokenize Method
+- Support for Continue on Error in Insert Method
+
+## [1.7.2] - 2023-08-28
+### Added
+-  Support for OFF Loglevel.
+
+## [1.7.1] - 2023-08-22
+### Changed
+-  Internal Batch API with tokenization
+
+## [1.7.0] - 2023-08-18
+### Added
+- Support for BYOT tokens in insert method
+- Support for Context in insert method
+
 ## [1.6.0] - 2023-06-09
 ### Added
 - `redaction` key for detokenize method for column group support.